requirements.txt
fastAPI
uvicorn
line-bot-sdk
python-dotenv
openai==0.19.0